Moved _radius and _footprintRadius from double to int and removed the sqrt() calls
Former-commit-id: 6ade3d39e052986b07dd9fd8c0f18fb95feeeab9

+++ roborobo3/include/core/World/CircleObject.h
@@ -21,8 +21,8 @@ class CircleObject : public PhysicalObject
 
 protected:
     
-    double _radius; // radius. In pixels.
-    double _footprintRadius; // radius of footprint, accessible to ground sensors. In pixels.
+    int _radius; // radius. In pixels.
+    int _footprintRadius; // radius of footprint, accessible to ground sensors. In pixels.
     
     // the impulses given to the object by nearby robots in the current time step, or other objects in the former
     std::map<int, Point2d> _impulses;
